+++ app/templates/coins.html
